requirements.txt
# pip install -r requirements.txt
# pip install --upgrade -r requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
openai
tiktoken
python-dotenv